Get questions by type from the survey sheet

get_survey_type() is a generic; the other functions get the specified obvious type from the survey sheet.

Usage

get_survey_type(survey, question_type)

get_survey_select_one(survey)

get_survey_select_multiple(survey)

get_survey_calculate(survey)

get_survey_integer(survey)

get_survey_decimal(survey)

Arguments

survey

(df[1+, (type = character, name = character, ...)]) The survey sheet from Kobo with column "type" split already, see ?split_survey(). If label_survey is TRUE, survey must also contain a "label" column. See split_survey()

question_type

(character[1]) Usually "integer", "select_one" or "select_multiple".

Value

(“) A character vector of select_one questions.

Details

survey should have a split type column with types of variables such as "select_one", "select_multiple", etc.