Kobo survey analysis from a data analysis plan

kobo_analysis_from_dap performs a survey analysis from a data analysis plan, while kobo_analysis_from_dap_group is a wrapper around the former to run multiple grouped analyses at once.

Usage

kobo_analysis_from_dap(
  design,
  dap,
  survey,
  choices,
  group = NULL,
  level = 0.95,
  choices_sep = "/"
)

kobo_analysis_from_dap_group(
  design,
  dap,
  survey,
  choices,
  l_group,
  no_group = TRUE,
  level = 0.95,
  choices_sep = "/"
)

Arguments

design

(srvyr::design) A srvyr::design object.

dap

(df) A well formatted data analysis plan.

survey

(df[1+, (type = character, name = character, ...)]) The survey sheet from Kobo with column "type" split already, see ?split_survey(). If label_survey is TRUE, survey must also contain a "label" column. See split_survey().

choices

(NULL | df[1+, (label = character, name = character, ...)]) The choices sheet from Kobo. If not NULL, the function tries to retrieve labels for choices.

group

(character[1+] | NULL) Variables to group by. (default: NULL for no group)

level

(For vartype = "ci" only) A single number or vector of numbers indicating the confidence level

choices_sep

(character[1]) Separator used in the Kobo extracted variable names for select_multiple questions. Default to "/".

l_group

(list[character]) A list of vectors of variables to group by.

no_group

(logical[1]) If TRUE, the analysis without grouping is run. (default: TRUE)

Value

(“) A survey-summarized-median data frame

Specifics

A well formatted data analysis plan is a data frame with at least three columns: analysis, vars, na_rm, where,

analysis is one of the following types: "mean", "median", "select_multiple", "select_one", "ratio", "interact".

vars are the variables to analyze. For "mean", "median", "select_multiple", and "select_one", it is a single variable. For "ratio", it is two variables separated by a comma, where the first one is the numerator and the second one the denominator. For "interact", it is two or more variables separated by commas, which will be interacted.

na_rm is either "yes" or "no", indicating whether NAs should be removed for that analysis.

It takes the default for most variables, e.g. it tries to label, it calculates confidence intervals, and separators are the default.

Choices cannot be NULL.

The separator for choices can be changed depending on how your dataset's variable names are formatted.

Level can be changed.

See also