Get all numeric variables's means

Get all numeric variables's means

Usage

kobo_mean_all(
  design,
  survey,
  group = NULL,
  group_key_sep = " -/- ",
  label_survey = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95
)

Arguments

design

(srvyr::design) A srvyr::design object.

survey

(df[1+, (type = character, name = character, ...)]) The survey sheet from Kobo with column "type" split already, see ?split_survey(). If label_survey is TRUE, survey must also contain a "label" column. See split_survey().

group

(character[1+] | NULL) Variables to group by. (default: NULL for no group)

group_key_sep

(character[1]) The string to separate grouping variables in a fancy 'group_key' column. (default: " -/- ")

label_survey

(logical[1]) Whether to retrieve labels of vars from the survey sheet. Default to TRUE.

na_rm

(logical[1]) Whether NAs in 'var' should be removed. (default: TRUE)

vartype

Report variability as one or more of: standard error ("se", default), confidence interval ("ci"), variance ("var") or coefficient of variation ("cv").

level

(For vartype = "ci" only) A single number or vector of numbers indicating the confidence level

Value

(“) A character vector of select_one questions.

Details

survey should have a split type column with types of variables such as "select_one", "select_multiple", etc.

See also