Check common parameters for svy_ functions

This helper function consolidates the common parameter validation patterns used across the svy_ family of functions to reduce code duplication.

Usage

svy_param_check(
  design,
  vars = NULL,
  nums = NULL,
  denoms = NULL,
  group = NULL,
  group_key_sep = NULL,
  na_rm = NULL,
  vartype = NULL,
  level = NULL,
  check_vars_not_in_group = TRUE,
  check_equal_length = TRUE,
  warn_low_ci = TRUE,
  separators = NULL
)

Arguments

design

(srvyr::design) A srvyr::design object.

vars

(character[1+] | NULL) Variables to analyze. Can be NULL for functions that don't use vars.

nums

(character[1+] | NULL) Numerator variables (for ratio functions). Can be NULL.

denoms

(character[1+] | NULL) Denominator variables (for ratio functions). Can be NULL.

group

(character[1+] | NULL) Variables to group by. Default to NULL for no group.

group_key_sep

(character[1] | NULL) The string to separate grouping variables. Can be NULL to skip check.

na_rm

(logical[1] | NULL) Should NAs be removed? Can be NULL to skip check.

vartype

(character[1] | NULL) Type of variance estimation. Can be NULL to skip check.

level

(numeric[1] | NULL) Confidence level between 0 and 1. Can be NULL to skip check.

check_vars_not_in_group

(logical[1]) Whether to check that vars are not in group (default: TRUE).

check_equal_length

(logical[1]) Whether to check that nums and denoms have equal length (default: TRUE).

warn_low_ci

(logical[1]) Whether to warn when CI level is below 90% (default: TRUE).

separators

(character[1+] | NULL) Additional separator parameters to validate. Can be NULL.

Value

(TRUE | error) TRUE if all validations pass, otherwise throws an error.